+++ pypdfbox/pdmodel/interactive/documentnavigation/destination/pd_page_fit_height_destination.py
@@ -23,6 +23,7 @@ class PDPageFitHeightDestination(PDPageDestination):
     def __init__(self, array: COSArray | None = None) -> None:
         super().__init__(array)
         if array is None:
+            self._array.grow_to_size(self._SLOT_LEFT + 1, COSNull.NULL)
             self._set_type(self.TYPE)
 
     def fit_bounding_box(self) -> bool: